Almeric First· push 20230417 (#2)
* GPT generated first blog

This is the blog model given by GPT4. should be able to add new posts on the right side of the blog. Need to use prompt to improve the quality. Not verified by google.

* Create README.md

The first readme

* GPT_changed1

The url is wrong. Changed according to GPT advice

* GPT Change #2

The second GPT web design, japanese Zen Style Added.

* Google Metadata added

To add my blog to google search engine

* The new free blog theme 20230417

Try if I can use it