Locally Jekyll Tested!
The first time the website is tested by Local Jekyll. Problkem is page is  default Jekyll and not needed.

It is a static blog generator that can generate powerful blogs. Before I don't know how to verify it locally but I correctly install Ruby by rbenv not using MacOs Ruby.